Clean up to message digest tests
Simplifies the coupling between base class and sub classes by removing
the need for sub classes to call a method from the parent class to
correctly initialise the message digest for testing. The sub classes are
not only required to implement a property to get the message digest to
be tested.